+++ src/neon/azext_neon/aaz/latest/neon/postgres/__init__.py
@@ -10,3 +10,4 @@
 
 from .__cmd_group import *
 from ._create import *
+from ._get_postgres_version import *